Smashed player objects with users, moved serializers into new module, cleaned up code.
