+++ frolferapi/serializers/pin.py
@@ -1,0 +1,19 @@
+from rest_framework import serializers
+from frolferapi.models import Pin
+
+class PinSerializer(serializers.ModelSerializer):
+    """JSON serializer for disc golf course pins.
+
+    Arguments:
+        serializers
+    """
+    class Meta:
+        model = Pin
+        fields = (
+            'id',
+            'label',
+            'par',
+            'length',
+            'hole_id',
+            'active'
+        )
