+++ frolferapi/serializers/player.py
@@ -1,0 +1,20 @@
+from rest_framework import serializers
+from frolferapi.serializers import UserSerializer
+
+from frolferapi.models import Player
+
+class PlayerSerializer(serializers.ModelSerializer):
+    """JSON serializer for disc golf players.
+
+    Arguments:
+        serializers
+    """
+
+    user = UserSerializer()
+    class Meta:
+        model = Player
+        fields = (
+            'id',
+            'user',
+            'rounds_played'
+        )
