+++ frolferapi/views/__init__.py
@@ -8,4 +8,4 @@ from .player import Players
 from .score_card import ScoreCards
 from .auth import login_user
 from .auth import register_user
-
+from .auth import logout_user
