+++ frolferapi/views/auth.py
@@ -1,12 +1,12 @@
 """Authentication viewset module."""
 import json
 from django.http import HttpResponse
-from django.contrib.auth import login, authenticate
+from django.contrib.auth import login, authenticate, logout
 from django.contrib.auth.models import User
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.authtoken.models import Token
-from rest_framework import serializers
 from frolferapi.models import Player
+from frolferapi.serializers import UserSerializer
 
 @csrf_exempt
 def login_user(request):
@@ -27,12 +27,22 @@ def login_user(request):
         if authenticated_user is not None:
             token = Token.objects.get(user=authenticated_user)
             serializer = UserSerializer(authenticated_user, context={ 'request': request })
+            login(request, authenticated_user)
             data = json.dumps({"valid": True, "token": token.key, "user": serializer.data})
             return HttpResponse(data, content_type='application/json')
 
         else:
             data = json.dumps({"valid": False})
             return HttpResponse(data, content_type='application/json')
+
+def logout_user(request):
+    '''Handles logging the user out.
+
+    Method arguments:
+      request -- The full HTTP request object
+    '''
+
+    logout(request)
 
 @csrf_exempt
 def register_user(request):
@@ -62,19 +72,3 @@ def register_user(request):
     serializer = UserSerializer(new_user, context={ 'request': request })
     data = json.dumps({"valid": True, "token": token.key, "user": serializer.data})
     return HttpResponse(data, content_type='application/json')
-
-class UserSerializer(serializers.ModelSerializer):
-    """
-    User Serializer
-    """
-    class Meta:
-        model = User
-        fields = (
-            'id',
-            'username',
-            'first_name',
-            'last_name',
-            'email',
-            'date_joined',
-            'is_staff'
-        )
